+++ aanalytics2/__init__.py
@@ -5,3 +5,4 @@ from .aanalytics14 import *
 from .configs import *
 from .projects import *
 from .requestCreator import *
+from .workspaceCreator import WorkspaceCreator, TextBuilder